requirements.txt
    tensorflow
    Pillow
    numpy
    flask
    gunicorn
    